pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "canvas-engineering"
version = "0.2.0"
description = "Prompt engineering, but for latent space. A type system for multimodal latent dynamics in video diffusion transformers."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    {name = "Jacob"},
    {name = "Claude Opus 4.6", email = "noreply@anthropic.com"},
]
keywords = ["diffusion", "transformer", "looped-attention", "video", "robotics", "canvas"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0",
    "einops>=0.7",
]

[project.optional-dependencies]
cogvideox = ["diffusers>=0.28", "transformers>=4.40", "accelerate"]
data = ["av", "numpy", "Pillow"]
dev = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/JacobFV/canvas-engineering"

[tool.setuptools.packages.find]
include = ["canvas_engineering*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
